extract annotateLines from LearningGoals
